Fix creation of ROOT dictionary
The CMake macro ROOT_GENERATE_DICTIONARY inherited from the used ROOT installtion
creates a dictionary with full paths to the header files located in the source directory.
This results in an unusable package when the package is installed and the source directory
is deleted. Remove the full path from the include directives to move the info to a
relative path. Using ROOT_INCLUDE_PATH at runtime then allows to find the header files.